Moved endpoint related utilities to helpers module
